refactor: Split pluginImplementation logic
