+++ packages/rownd-nodejs/src/plugin.ts
@@ -6,16 +6,33 @@ import { createPluginInitFunction } from "@shared/js";
 import { withRequestHandler } from "@shared/nodejs";
 import { createInstance } from "@rownd/node";
 import supertokens from "supertokens-node";
-import { HANDLE_BASE_PATH, PLUGIN_ID, PLUGIN_SDK_VERSION } from "./constants";
+import {
+  HANDLE_BASE_PATH,
+  HUB_VERIFY_EMAIL_PAGE_PATH,
+  HUB_LOGIN_PAGE_PATH,
+  PLUGIN_ID,
+  PLUGIN_SDK_VERSION,
+} from "./constants";
 import { RowndPluginConfig, RowndPluginNormalisedConfig } from "./types";
 import { enableDebugLogs, logDebugMessage } from "./logger";
 import { createClient } from "./telemetry/createTelemetryClient";
+import { assertRowndAppVariantIsConfigured, setPluginConfig } from "./config";
+import { shouldLinkRowndAccounts } from "./rownd-compatibility";
+import { setRowndClient } from "./rownd-repository";
 import {
-  setRowndClient,
-  setPluginConfig,
   buildRowndSessionClaims,
   completePendingEmailVerification,
+  recordRowndAppVariantForUser,
   RowndIsAnonymousClaim,
+} from "./supertokens-repository";
+import {
+  getRequestedAppVariantIdFromRequest,
+  getRequestedDisplayContextFromRequest,
+  getRequestedRedirectToPathFromRequest,
+  rewriteLinkPath,
+  rewriteLinkToMobileDeepLink,
+} from "./utils";
+import {
   handleDeleteUser,
   handleGetAppConfig,
   handleGetUser,
@@ -27,14 +44,6 @@ import {
   handleUpdateUser,
   handleUpdateUserField,
   handleUpdateUserMeta,
-  shouldLinkRowndAccounts,
-  rewriteLinkPath,
-  getRequestedAppVariantIdFromRequest,
-  getRequestedDisplayContextFromRequest,
-  getRequestedRedirectToPathFromRequest,
-  recordRowndAppVariantForUser,
-  assertRowndAppVariantIsConfigured,
-  rewriteLinkToMobileDeepLink,
 } from "./pluginImplementation";
 
 export const init: (config: RowndPluginConfig) => SuperTokensPlugin =
@@ -235,13 +244,13 @@ export const init: (config: RowndPluginConfig) => SuperTokensPlugin =
                     return {
                       ...implementation,
                       sendEmail: async function (input) {
-                        return implementation.sendEmail({
-                          ...addHubBootstrapParams(
+                        return implementation.sendEmail(
+                          addHubBootstrapParams(
                             input,
                             "urlWithLinkCode",
-                            "account/login",
+                            HUB_LOGIN_PAGE_PATH,
                           ),
-                        });
+                        );
                       },
                     };
                   },
@@ -259,13 +268,13 @@ export const init: (config: RowndPluginConfig) => SuperTokensPlugin =
                     return {
                       ...implementation,
                       sendSms: async function (input) {
-                        return implementation.sendSms({
-                          ...addHubBootstrapParams(
+                        return implementation.sendSms(
+                          addHubBootstrapParams(
                             input,
                             "urlWithLinkCode",
-                            "account/login",
+                            HUB_LOGIN_PAGE_PATH,
                           ),
-                        });
+                        );
                       },
                     };
                   },
@@ -286,32 +295,24 @@ export const init: (config: RowndPluginConfig) => SuperTokensPlugin =
                 const displayContext = getRequestedDisplayContextFromRequest(
                   input.options.req,
                 );
+                if (displayContext) {
+                  input.userContext.rowndDisplayContext = displayContext;
+                }
                 const redirectToPath = getRequestedRedirectToPathFromRequest(
                   input.options.req,
                 );
+                if (redirectToPath) {
+                  input.userContext.rowndRedirectToPath = redirectToPath;
+                }
                 const appVariantId = getRequestedAppVariantIdFromRequest(
                   input.options.req,
                 );
+                if (appVariantId) {
+                  input.userContext.rowndAppVariantId = appVariantId;
+                }
                 assertRowndAppVariantIsConfigured(appVariantId);
 
-                return originalImplementation.createCodePOST({
-                  ...input,
-                  userContext:
-                    displayContext || redirectToPath || appVariantId
-                      ? {
-                          ...input.userContext,
-                          ...(appVariantId
-                            ? { rowndAppVariantId: appVariantId }
-                            : {}),
-                          ...(displayContext
-                            ? { rowndDisplayContext: displayContext }
-                            : {}),
-                          ...(redirectToPath
-                            ? { rowndRedirectToPath: redirectToPath }
-                            : {}),
-                        }
-                      : input.userContext,
-                });
+                return originalImplementation.createCodePOST(input);
               },
               consumeCodePOST: async (
                 input: Parameters<
@@ -327,13 +328,12 @@ export const init: (config: RowndPluginConfig) => SuperTokensPlugin =
                 const appVariantId = getRequestedAppVariantIdFromRequest(
                   input.options.req,
                 );
+                if (appVariantId) {
+                  input.userContext.rowndAppVariantId = appVariantId;
+                }
                 assertRowndAppVariantIsConfigured(appVariantId);
-                const response = await originalImplementation.consumeCodePOST({
-                  ...input,
-                  userContext: appVariantId
-                    ? { ...input.userContext, rowndAppVariantId: appVariantId }
-                    : input.userContext,
-                });
+                const response =
+                  await originalImplementation.consumeCodePOST(input);
 
                 if (response.status === "OK") {
                   await recordRowndAppVariantForUser(
@@ -361,13 +361,12 @@ export const init: (config: RowndPluginConfig) => SuperTokensPlugin =
                 const appVariantId = getRequestedAppVariantIdFromRequest(
                   input.options.req,
                 );
+                if (appVariantId) {
+                  input.userContext.rowndAppVariantId = appVariantId;
+                }
                 assertRowndAppVariantIsConfigured(appVariantId);
-                const response = await originalImplementation.signInUpPOST({
-                  ...input,
-                  userContext: appVariantId
-                    ? { ...input.userContext, rowndAppVariantId: appVariantId }
-                    : input.userContext,
-                });
+                const response =
+                  await originalImplementation.signInUpPOST(input);
 
                 if (response.status === "OK") {
                   await recordRowndAppVariantForUser(
@@ -412,19 +411,24 @@ export const init: (config: RowndPluginConfig) => SuperTokensPlugin =
             functions: (originalImplementation) => ({
               ...originalImplementation,
               createNewSession: async (input) => {
+                const [rowndSessionClaims, rowndIsAnonymousClaim] =
+                  await Promise.all([
+                    buildRowndSessionClaims(
+                      input.userId,
+                      input.accessTokenPayload,
+                    ),
+                    RowndIsAnonymousClaim.build(
+                      input.userId,
+                      input.recipeUserId,
+                      input.tenantId,
+                      input.accessTokenPayload,
+                      input.userContext,
+                    ),
+                  ]);
                 input.accessTokenPayload = {
                   ...input.accessTokenPayload,
-                  ...(await buildRowndSessionClaims(
-                    input.userId,
-                    input.accessTokenPayload,
-                  )),
-                  ...(await RowndIsAnonymousClaim.build(
-                    input.userId,
-                    input.recipeUserId,
-                    input.tenantId,
-                    input.accessTokenPayload,
-                    input.userContext,
-                  )),
+                  ...rowndSessionClaims,
+                  ...rowndIsAnonymousClaim,
                 };
 
                 return originalImplementation.createNewSession(input);
@@ -456,7 +460,7 @@ export const init: (config: RowndPluginConfig) => SuperTokensPlugin =
                           ...addHubBootstrapParams(
                             input,
                             "emailVerifyLink",
-                            "account/verify-email",
+                            HUB_VERIFY_EMAIL_PAGE_PATH,
                           ),
                         });
                       },
